Additional functionality + code styling fixes
Ok, now merging to develop and later will do deeper checks

+++ mltu/dataProvider.py
@@ -9,7 +9,8 @@ from .augmentors import Augmentor
 from .transformers import Transformer
 
 import logging
-logging.basicConfig(format='%(asctime)s %(levelname)s %(name)s: %(message)s')
+logging.basicConfig(format="%(asctime)s %(levelname)s %(name)s: %(message)s")
+
 
 class DataProvider:
     def __init__(
@@ -238,7 +239,7 @@ class DataProvider:
 
         # Convert to numpy array if not already
         # TODO: This is a hack, need to fix this
-        if not isinstance(annotation, (np.ndarray, int, float, str, np.uint8, np.float)):
+        if not isinstance(annotation, (np.ndarray, int, float, str, np.uint8, float)):
             annotation = annotation.numpy()
 
         return data, annotation
